refactor(backend-user-domain): move user models
move models to distinguish entities and value objects
