+++ packages/backend/apps/user/backend-user-domain/src/users/domain/specs/UserCanCreateCodeSpec.ts
@@ -1,6 +1,6 @@
 import { Spec } from '@cornie-js/backend-common';
 
-import { User } from '../models/User';
+import { User } from '../entities/User';
 
 export class UserCanCreateCodeSpec implements Spec<[User]> {
   public isSatisfiedBy(user: User): boolean {
